Remove the arbitrary 100 to-device message limit on AS transactions
It turned out that this was harder to keep track of than originally thought. Because
multiple to-device messages can share the same stream id, counting the number of
to-device messages here doesn't make sense. You could end up only
sending 100 messages, but there's actually 150 to-device messages which
all share the same stream id.

If we were to reintroduce the limit, it should be a limit on the stream ids,
number of unique stream ids, rather than the count of individual messages.