Add tests
I decided to spin up another test class for this as the existing one is
1. quite old and 2. was mocking away too much of the infrastructure to
my liking. I've named the new class alluding to ephemeral messages, and
while we already have some ephemeral tests in AppServiceHandlerTestCase,
ideally I'd like to migrate those over.

There's two new tests here. One for testing that to-device messages for
a local user are received by any application services that have
registered interest in that user - and that those that haven't won't
receive those messages.

The next test is similar, but tests with a whole bunch of to-device
messages. Rather than registering tons of devices - which would make for
a very slow unit test - I decided to just directly insert most of the
to-device messages into the database, then send a final one via the API.
Sending via the API is necessary to kick off the appservice scheduler.
It also allows us to test old messages being sent out correctly.